Refactor method middleware
