+++ app/scripts/lib/enums.js
@@ -14,6 +14,7 @@ const MESSAGE_TYPE = {
   ETH_GET_ENCRYPTION_PUBLIC_KEY: 'eth_getEncryptionPublicKey',
   ETH_SIGN: 'eth_sign',
   ETH_SIGN_TYPED_DATA: 'eth_signTypedData',
+  LOG_WEB3_USAGE: 'metamask_logInjectedWeb3Usage',
   PERSONAL_SIGN: 'personal_sign',
 }
