+++ app/scripts/lib/rpc-method-middleware/handlers/index.js
@@ -1,0 +1,6 @@
+import logWeb3Usage from './log-web3-usage'
+
+const handlers = [
+  logWeb3Usage,
+]
+export default handlers
